# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "this is the swagger implementation for dialosoft",
        "title": "Dialosoft Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "@Flussen on github",
            "email": "flussen0@gmail.com"
        },
        "license": {
            "name": "GPL-3.0 License",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "0.2.0"
    },
    "host": "localhost:8080",
    "basePath": "/dialosoft-api/v1",
    "paths": {
...
        "/forums/get-forums-by-category-id/{categoryID}": {
            "get": {
                "security": [
                    {
                        "AccessTokenAuth": []
                    },
                    {
                        "RefreshTokenAuth": []
                    }
                ],
                "description": "Retrieves forums by category ID if allowed for the role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "Get forums by category ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ForumDto"
                            }
                        }
                    },
                    "400": {
                        "description": "BAD REQUEST",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "403": {
                        "description": "FORBIDDEN",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    }
                }
            }
        },
        "/forums/protected/create-new-forum": {
            "post": {
                "security": [
                    {
                        "AccessTokenAuth": []
                    },
                    {
                        "RefreshTokenAuth": []
                    }
                ],
                "description": "Creates a new forum with the specified details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "Create a new forum",
                "parameters": [
                    {
                        "description": "New Forum Data",
                        "name": "forum",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.NewForum"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "BAD REQUEST",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "409": {
                        "description": "CONFLICT",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    }
                }
            }
        },
        "/forums/protected/delete-forum/{id}": {
            "delete": {
                "security": [
                    {
                        "AccessTokenAuth": []
                    },
                    {
                        "RefreshTokenAuth": []
                    }
                ],
                "description": "Deletes a forum with the specified ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "Delete a forum",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Forum ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DELETED",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "BAD REQUEST",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    }
                }
            }
        },
        "/forums/protected/get-all-forums": {
            "get": {
                "security": [
                    {
                        "AccessTokenAuth": []
                    },
                    {
                        "RefreshTokenAuth": []
                    }
                ],
                "description": "Retrieves a list of all forums.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "Get all forums",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ForumDto"
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    }
                }
            }
        },
        "/forums/protected/get-forum-by-id/{id}": {
            "get": {
                "security": [
                    {
                        "AccessTokenAuth": []
                    },
                    {
                        "RefreshTokenAuth": []
                    }
                ],
                "description": "Retrieves a specific forum by its unique ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "Get forum by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Forum ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ForumDto"
                        }
                    },
                    "400": {
                        "description": "BAD REQUEST",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    }
                }
            }
        },
        "/forums/protected/get-forum-by-name/{name}": {
            "get": {
                "security": [
                    {
                        "AccessTokenAuth": []
                    },
                    {
                        "RefreshTokenAuth": []
                    }
                ],
                "description": "Retrieves a specific forum by its unique name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "Get forum by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Forum Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ForumDto"
                        }
                    },
                    "400": {
                        "description": "BAD REQUEST",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    }
                }
            }
        },
        "/forums/protected/restore-forum/{id}": {
            "put": {
                "security": [
                    {
                        "AccessTokenAuth": []
                    },
                    {
                        "RefreshTokenAuth": []
                    }
                ],
                "description": "Restores a deleted forum with the specified ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "Restore a forum",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Forum ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "RESTORED",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "BAD REQUEST",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    }
                }
            }
        },
        "/forums/protected/update-forum/{id}": {
            "put": {
                "security": [
                    {
                        "AccessTokenAuth": []
                    },
                    {
                        "RefreshTokenAuth": []
                    }
                ],
                "description": "Updates an existing forum with the specified ID and details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "Update a forum",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Forum ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Forum Data",
                        "name": "forum",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.NewForum"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UPDATED",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "BAD REQUEST",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.ForumDto": {
            "type": "object",
            "required": [
                "categoryId",
                "name",
                "type"
            ],
            "properties": {
                "categoryId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "rolesAllowed": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "request.NewForum": {
            "type": "object",
            "required": [
                "categoryID"
            ],
            "properties": {
                "categoryID": {
                    "type": "string",
                    "example": "8213280e-2000-403a-b375-cdcda6488450"
                },
                "description": {
                    "type": "string",
                    "example": "some description for forum 2"
                },
                "isActive": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Forum 2"
                },
                "rolesAllowed": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "['8213280e-2000-403a-b375-cdcda6488450'",
                        " 'roleUUID2']"
                    ]
                },
                "type": {
                    "type": "string",
                    "example": "information"
                }
            }
        },
...
        "response.StandardError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "An error occurred - some context"
                }
            }
        },
        "response.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "info message"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AccessTokenAuth": {
            "description": "Bearer token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "RefreshTokenAuth": {
            "description": "Refresh token for authentication",
            "type": "apiKey",
            "name": "X-Refresh-Token",
            "in": "header"
        }
    }
}